feat(wpm): add restart strategies
This commit adds an initial set of restart strategies: Never, Always and
OnFailure. OnFailure only triggers if a process was externally
terminated with a non-success exit code.

Serialization of unit definitions has been updated to use PascalCase to
hopefully be more familiar to systemd users and users familiar with INI
files.