feat(wpm): add forking and proc liveness healthcheck
This commit adds an initial implementation of a ServiceKind::Forking and
a Healthcheck::Process, which should allow users to run cmd-a.exe which
exits while launching and persisting cmd-b.exe in the background.

A process liveness healthcheck can now be added on the forked cmd-b.exe.
This is still very messy and there is a lot of code duplication which
needs to be cleaned up.