refactor(wpm): block thread for oneshot + forking services
This commit ensures that the main thread will block for Oneshot and
Forking services.

Already, a common use case for forking services is to start an elevated
process using tools like gsudo, which requires user interaction to
confirm process elevation - this should be on the main thread rather
than a background thread to ensure that the healthchecks only run after
the user has granted permissions for the ExecStart command to proceed as
expected.

ExecPostStart commands for Forking services will now also run only after
a successful healthcheck.