refactor(wpmd): defer socket message processing
This commit introduces a channel to enqueue socket messages as they are
received from wpmctl.

As a result of this, their processing can be deferred and we never have
to take a lock on ProcessManager in the socket listener's hot path.

Messages on the rx side of the channel are processed in a separate
thread where a potentially long-lived mutex lock will not result in the
socket listener being poisoned as it was previously.