feat(wpm): add remote executable fetching
This commit adds some basic package management and pinning abilities to
wpm.

In addition to an executable being given by the user as a local file
(deserialized to a PathBuf), a user can now also specify a remote URL +
sha256 hash as an executable, or a Scoop package manifest remote URL for
a pinned version of a package.

In the case of the remote URL + sha256, wpmd will check if the file
exists in the wpm store, and if it does, it will use this as a regular
local executable file, and if it doesn't, it will download the file,
verify against the sha256 hash, and then save it in the wpm store for
future use.

In the case of a Scoop package manifest, wpmd will check if the pinned
version of the package is available in scoop/apps, and if it is, it will
use the target (which defaults to {package_name}.exe, but can be
configured by the user for multi-binary packages) as the local
executable file. If it isn't, `scoop install` will be called on the
remote manifest URL.

Users can run `wpmctl rebuild` ensure all remote executables and
packages exist on the filesystem before launching `wpmd`.